Restructure app and initialize product to hardcoded id
