+++ client/components/App.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
 
 const App = () => (
-  <div>
+  <>
+    {/* place main module components within this fragment */}
     <h1>7 Satsumas</h1>
-  </div>
+  </>
 );
 
 export default App;
