+++ client/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import App from './components/App';
-import { ProductProvider } from './components/ProductContext';
+import { ProductProvider } from './ProductContext';
 
 render(
   <ProductProvider>
